[NEW RBO] Cleaned up expression extract rule, moved methods to appropriate files (#33656)
