+++ ydb/core/kqp/opt/rbo/kqp_operator.cpp
@@ -3,253 +3,11 @@
 #include "kqp_rbo_utils.h"
 #include <yql/essentials/core/yql_expr_optimize.h>
 
-namespace {
-using namespace NKikimr;
-using namespace NKqp;
-using namespace NYql;
-using namespace NNodes;
-
-void DFS(int vertex, TVector<int> &sortedStages, THashSet<int> &visited, const THashMap<int, TVector<int>> &stageInputs) {
-    visited.emplace(vertex);
-
-    for (auto u : stageInputs.at(vertex)) {
-        if (!visited.contains(u)) {
-            DFS(u, sortedStages, visited, stageInputs);
-        }
-    }
-
-    sortedStages.push_back(vertex);
-}
-
-TExprNode::TPtr AddRenames(TExprNode::TPtr input, const TStageGraph::TSourceStageTraits& sourceStagesTraits, TExprContext& ctx) {
-    if (sourceStagesTraits.Renames.empty()) {
-        return input;
-    }
-
-    TVector<TExprBase> items;
-    auto arg = Build<TCoArgument>(ctx, input->Pos()).Name("arg").Done().Ptr();
-
-    for (const auto& rename : sourceStagesTraits.Renames) {
-        // clang-format off
-        auto tuple = Build<TCoNameValueTuple>(ctx, input->Pos())
-            .Name().Build(rename.second.GetFullName())
-            .Value<TCoMember>()
-                .Struct(arg)
-                .Name().Build(rename.first)
-            .Build()
-        .Done();
-        // clang-format on
-        items.push_back(tuple);
-    }
-
-    // clang-format off
-    return Build<TCoMap>(ctx, input->Pos())
-        .Input(input)
-        .Lambda<TCoLambda>()
-            .Args({arg})
-            .Body<TCoAsStruct>()
-                .Add(items)
-            .Build()
-        .Build()
-    .Done().Ptr();
-    // clang-format on
-}
-
-TExprNode::TPtr BuildSourceStage(TExprNode::TPtr dqsource, TExprContext &ctx) {
-    auto arg = Build<TCoArgument>(ctx, dqsource->Pos()).Name("arg").Done().Ptr();
-    // clang-format off
-    return Build<TDqPhyStage>(ctx, dqsource->Pos())
-        .Inputs()
-            .Add({dqsource})
-        .Build()
-        .Program()
-            .Args({arg})
-            .Body(arg)
-        .Build()
-        .Settings().Build()
-    .Done().Ptr();
-    // clang-format on
-}
-
-} // namespace
-
 namespace NKikimr {
 namespace NKqp {
 
-template <class T>
-void AddUnique(TVector<T>& toAdd, TVector<T>& target) {
-    for (const auto & e : toAdd) {
-        if (std::find(target.begin(), target.end(), e) == target.end()) {
-            target.push_back(e);
-        }
-    }
-}
-
 using namespace NYql;
 using namespace NNodes;
-
-template <typename DqConnectionType>
-TExprNode::TPtr TConnection::BuildConnectionImpl(TExprNode::TPtr inputStage, TPositionHandle pos, TExprNode::TPtr& newStage, TExprContext& ctx) {
-    if (FromSourceStageStorageType == NYql::EStorageType::RowStorage) {
-        inputStage = BuildSourceStage(inputStage, ctx);
-        newStage = inputStage;
-    }
-    // clang-format off
-    return Build<DqConnectionType>(ctx, pos)
-        .Output()
-            .Stage(inputStage)
-            .Index().Build(ToString(OutputIndex))
-        .Build()
-    .Done().Ptr();
-    // clang-format on
-}
-
-TExprNode::TPtr TBroadcastConnection::BuildConnection(TExprNode::TPtr inputStage, TPositionHandle pos, TExprNode::TPtr& newStage, TExprContext& ctx) {
-    return BuildConnectionImpl<TDqCnBroadcast>(inputStage, pos, newStage, ctx);
-}
-
-TExprNode::TPtr TMapConnection::BuildConnection(TExprNode::TPtr inputStage, TPositionHandle pos, TExprNode::TPtr& newStage, TExprContext& ctx) {
-    return BuildConnectionImpl<TDqCnMap>(inputStage, pos, newStage, ctx);
-}
-
-TExprNode::TPtr TUnionAllConnection::BuildConnection(TExprNode::TPtr inputStage, TPositionHandle pos, TExprNode::TPtr &newStage,
-                                                     TExprContext &ctx) {
-    return BuildConnectionImpl<TDqCnUnionAll>(inputStage, pos, newStage, ctx);
-}
-
-TExprNode::TPtr TShuffleConnection::BuildConnection(TExprNode::TPtr inputStage, TPositionHandle pos, TExprNode::TPtr& newStage,
-                                                    TExprContext& ctx) {
-    if (FromSourceStageStorageType == NYql::EStorageType::RowStorage) {
-        inputStage = BuildSourceStage(inputStage, ctx);
-        newStage = inputStage;
-    }
-
-    TVector<TCoAtom> keyColumns;
-    for (const auto& k : Keys) {
-        TString columnName;
-        if (FromSourceStageStorageType != NYql::EStorageType::NA || k.GetAlias() == "") {
-            columnName = k.GetColumnName();
-        } else {
-            columnName = k.GetFullName();
-        }
-        auto atom = Build<TCoAtom>(ctx, pos).Value(columnName).Done();
-        keyColumns.push_back(atom);
-    }
-
-    // clang-format off
-    return Build<TDqCnHashShuffle>(ctx, pos)
-        .Output()
-            .Stage(inputStage)
-            .Index().Build(ToString(OutputIndex))
-        .Build()
-        .KeyColumns()
-            .Add(keyColumns)
-        .Build()
-    .Done().Ptr();
-    // clang-format on
-}
-
-TExprNode::TPtr TMergeConnection::BuildConnection(TExprNode::TPtr inputStage, TPositionHandle pos, TExprNode::TPtr& newStage,
-                                                  TExprContext& ctx) {
-    if (FromSourceStageStorageType == NYql::EStorageType::RowStorage) {
-        inputStage = BuildSourceStage(inputStage, ctx);
-        newStage = inputStage;
-    }
-
-    TVector<TExprNode::TPtr> sortColumns;
-    for (const auto& sortElement : Order) {
-        // clang-format off
-        sortColumns.push_back(Build<TDqSortColumn>(ctx, pos)
-            .Column<TCoAtom>().Build(sortElement.SortColumn.GetFullName())
-            .SortDirection().Build(sortElement.Ascending ? TTopSortSettings::AscendingSort : TTopSortSettings::DescendingSort)
-            .Done().Ptr());
-        // clang-format on
-    }
-
-    // clang-format off
-    return Build<TDqCnMerge>(ctx, pos)
-        .Output()
-            .Stage(inputStage)
-            .Index().Build(ToString(OutputIndex))
-        .Build()
-        .SortColumns()
-            .Add(sortColumns)
-        .Build()
-    .Done().Ptr();
-    // clang-format on
-}
-
-TExprNode::TPtr TSourceConnection::BuildConnection(TExprNode::TPtr inputStage, TPositionHandle pos, TExprNode::TPtr &newStage,
-                                                   TExprContext &ctx) {
-    Y_UNUSED(pos);
-    Y_UNUSED(newStage);
-    Y_UNUSED(ctx);
-    return inputStage;
-}
-
-std::pair<TExprNode::TPtr, TExprNode::TPtr> TStageGraph::GenerateStageInput(int& stageInputCounter, TExprNode::TPtr& node,
-                                                                            TExprContext& ctx, int fromStage) {
-    TString inputName = "input_arg" + std::to_string(stageInputCounter++);
-    YQL_CLOG(TRACE, CoreDq) << "Created stage argument " << inputName;
-    auto arg = Build<TCoArgument>(ctx, node->Pos()).Name(inputName).Done().Ptr();
-    auto output = arg;
-
-    if (IsSourceStage(fromStage)) {
-        output = AddRenames(arg, SourceStageRenames.at(fromStage), ctx);
-    }
-
-    return std::make_pair(arg, output);
-}
-
-void TStageGraph::TopologicalSort() {
-    TVector<int> sortedStages;
-    THashSet<int> visited;
-
-    for (auto id : StageIds) {
-        if (!visited.contains(id)) {
-            DFS(id, sortedStages, visited, StageInputs);
-        }
-    }
-
-    StageIds = sortedStages;
-}
-
-std::pair<TExprNode::TPtr, TVector<TExprNode::TPtr>> BuildSortKeySelector(const TVector<TSortElement>& sortElements, TExprContext& ctx, TPositionHandle pos) {
-    auto arg = Build<TCoArgument>(ctx, pos).Name("arg").Done().Ptr();
-    TVector<TExprNode::TPtr> directions;
-    TVector<TExprNode::TPtr> members;
-
-    for (const auto& element : sortElements) {
-        // clang-format off
-        members.push_back(Build<TCoMember>(ctx, pos)
-            .Struct(arg)
-            .Name().Build(element.SortColumn.GetFullName())
-            .Done().Ptr());
-        // clang-format on
-
-        directions.push_back(Build<TCoBool>(ctx, pos).Literal().Build(element.Ascending ? "true" : "false").Done().Ptr());
-    }
-
-    TExprNode::TPtr selector;
-    if (sortElements.size() == 1) {
-        // clang-format off
-        selector = Build<TCoLambda>(ctx, pos)
-            .Args({arg})
-            .Body(members[0])
-            .Done().Ptr();
-        // clang-format on
-    }
-    else {
-        // clang-format off
-        selector = Build<TCoLambda>(ctx, pos)
-            .Args({arg})
-            .Body<TExprList>().Add(members).Build()
-            .Done().Ptr();
-        // clang-format on
-    }
-
-    return std::make_pair(selector, directions);
-}
 
 /**
  * Base class Operator methods
@@ -1069,28 +827,5 @@ void TOpRoot::PlanToStringRec(std::shared_ptr<IOperator> op, TExprContext& ctx, 
     }
 }
 
-TVector<TInfoUnit> IUSetDiff(TVector<TInfoUnit> left, TVector<TInfoUnit> right) {
-    TVector<TInfoUnit> res;
-    for (const auto& unit : left) {
-        if (std::find(right.begin(), right.end(), unit) == right.end()) {
-            if (std::find(res.begin(), res.end(), unit) == res.end()) {
-                res.push_back(unit);
-            }
-        }
-    }
-    return res;
-}
-
-TVector<TInfoUnit> IUSetIntersect(TVector<TInfoUnit> left, TVector<TInfoUnit> right) {
-    TVector<TInfoUnit> res;
-    for (const auto& unit : left) {
-        if (std::find(right.begin(), right.end(), unit) != right.end()) {
-            if (std::find(res.begin(), res.end(), unit) == res.end()) {
-                res.push_back(unit);
-            }
-        }
-    }
-    return res;
-}
 } // namespace NKqp
 } // namespace NKikimr
