+++ ydb/core/kqp/opt/rbo/kqp_rbo_utils.cpp
@@ -1,0 +1,33 @@
+#include "kqp_rbo_utils.h"
+
+namespace NKikimr {
+namespace NKqp {
+
+using namespace NYql;
+
+TVector<TInfoUnit> IUSetDiff(TVector<TInfoUnit> left, TVector<TInfoUnit> right) {
+    TVector<TInfoUnit> res;
+    for (const auto& unit : left) {
+        if (std::find(right.begin(), right.end(), unit) == right.end()) {
+            if (std::find(res.begin(), res.end(), unit) == res.end()) {
+                res.push_back(unit);
+            }
+        }
+    }
+    return res;
+}
+
+TVector<TInfoUnit> IUSetIntersect(TVector<TInfoUnit> left, TVector<TInfoUnit> right) {
+    TVector<TInfoUnit> res;
+    for (const auto& unit : left) {
+        if (std::find(right.begin(), right.end(), unit) != right.end()) {
+            if (std::find(res.begin(), res.end(), unit) == res.end()) {
+                res.push_back(unit);
+            }
+        }
+    }
+    return res;
+}
+
+}
+}
