+++ ydb/core/kqp/opt/rbo/kqp_rbo_utils.h
@@ -12,7 +12,13 @@ using namespace NYql;
 
 TVector<TInfoUnit> IUSetDiff(TVector<TInfoUnit> left, TVector<TInfoUnit> right);
 TVector<TInfoUnit> IUSetIntersect(TVector<TInfoUnit> left, TVector<TInfoUnit> right);
-template <class T> void AddUnique(TVector<T>&, TVector<T>&);
+template <class T> void AddUnique(TVector<T>& toAdd, TVector<T>& target) {
+    for (const auto & e : toAdd) {
+        if (std::find(target.begin(), target.end(), e) == target.end()) {
+            target.push_back(e);
+        }
+    }
+}
 
 }
 }
